Consolidate various libwg implementations
