+++ wireguard/libwg/go.mod
@@ -1,0 +1,9 @@
+module github.com/mullvad/mullvadvpn-app/wireguard/libwg
+
+go 1.13
+
+require (
+	golang.org/x/sys v0.0.0-20200202164722-d101bd2416d5
+	golang.zx2c4.com/wireguard v0.0.20200121
+	golang.zx2c4.com/wireguard/windows v0.0.38
+)
