+++ wireguard/libwg/libwg.go
@@ -1,0 +1,69 @@
+/* SPDX-License-Identifier: Apache-2.0
+ *
+ * Copyright (C) 2017-2019 Jason A. Donenfeld <Jason@zx2c4.com>. All Rights Reserved.
+ * Copyright (C) 2020 Mullvad VPN AB. All Rights Reserved.
+ */
+
+package main
+
+// #include <stdlib.h>
+import "C"
+
+import (
+	"bufio"
+	"bytes"
+	"runtime"
+	"unsafe"
+	
+	"golang.zx2c4.com/wireguard/device"
+	"github.com/mullvad/mullvadvpn-app/wireguard/libwg/tunnelcontainer"
+)
+
+var tunnels tunnelcontainer.Container
+
+func init() {
+	device.RoamingDisabled = true
+	tunnels = tunnelcontainer.New()
+}
+
+//export wgTurnOff
+func wgTurnOff(tunnelHandle int32) {
+	{
+		tunnel, err := tunnels.Remove(tunnelHandle)
+		if err != nil {
+			return
+		}
+		tunnel.Device.Close()
+	}
+	// Calling twice convinces the GC to release NOW.
+	runtime.GC()
+	runtime.GC()
+}
+
+//export wgGetConfig
+func wgGetConfig(tunnelHandle int32) *C.char {
+	tunnel, err := tunnels.Get(tunnelHandle)
+	if err != nil {
+		return nil
+	}
+	settings := new(bytes.Buffer)
+	writer := bufio.NewWriter(settings)
+	if err := tunnel.Device.IpcGetOperation(writer); err != nil {
+		tunnel.Logger.Error.Println("Failed to get config for tunnel: ", err)
+		return nil
+	}
+	writer.Flush()
+	return C.CString(settings.String())
+}
+
+//export wgFreePtr
+func wgFreePtr(ptr unsafe.Pointer) {
+	C.free(ptr)
+}
+
+//export wgVersion
+func wgVersion() *C.char {
+	return C.CString(device.WireGuardGoVersion)
+}
+
+func main() {}
