--- wireguard/wireguard-go-windows/go.mod
@@ -1,21 +1,0 @@
-module github.com/mullvad/mullvadvpn-app/wireguard-go-windows
-
-go 1.12
-
-require (
-	github.com/Microsoft/go-winio v0.4.14 // indirect
-	github.com/konsorten/go-windows-terminal-sequences v1.0.2 // indirect
-	github.com/sirupsen/logrus v1.4.2 // indirect
-	github.com/stretchr/objx v0.2.0 // indirect
-	github.com/stretchr/testify v1.4.0 // indirect
-	golang.org/x/crypto v0.0.0-20200108215511-5d647ca15757 // indirect
-	golang.org/x/sync v0.0.0-20190911185100-cd5d95a43a6e // indirect
-	golang.org/x/sys v0.0.0-20200107162124-548cf772de50
-	golang.org/x/tools v0.0.0-20200108203644-89082a384178 // indirect
-	golang.org/x/xerrors v0.0.0-20191204190536-9bdfabe68543 // indirect
-	golang.zx2c4.com/wireguard v0.0.20200121
-	golang.zx2c4.com/wireguard/windows v0.0.38
-	gopkg.in/Knetic/govaluate.v3 v3.0.0 // indirect
-	gopkg.in/check.v1 v1.0.0-20190902080502-41f04d3bba15 // indirect
-	gopkg.in/yaml.v2 v2.2.7 // indirect
-)
