--- wireguard/wireguard-go/go.mod
@@ -1,8 +1,0 @@
-module github.com/mullvad/mullvadvpn-app/wireguard-go
-
-go 1.12
-
-require (
-	golang.org/x/sys v0.0.0-20191003212358-c178f38b412c
-	golang.zx2c4.com/wireguard v0.0.20200121
-)
